requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
bleach==6.2.0
Bottleneck @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_33wnmqdh9g/croot/bottleneck_1731058648266/work
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-markdownx==4.0.7
fonttools==4.56.0
frozendict==2.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.0
multitasking==0.0.11
numexpr @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/numexpr_1731706658483/work
numpy @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_1erks0t2_6/croot/numpy_and_numpy_base_1738078508172/work/dist/numpy-2.2.2-cp313-cp313-macosx_10_15_x86_64.whl#sha256=ecc5e8ab4fcd5231a55259cee0cd72d88322c0cc13648e72668afb95baf73a7e
openai==1.63.2
packaging==24.2
pandas @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_d6o0iypzho/croot/pandas_1732735102618/work/dist/pandas-2.2.3-cp313-cp313-macosx_10_15_x86_64.whl#sha256=2ef48dcabc48020c306097c5a1b12e25a7ed5e97ad5001b6be3e9b04c16756ef
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/python-dateutil_1728593996450/work
pytz @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pytz_1728595515863/work
requests==2.32.3
scipy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_06t04e6sdk/croot/scipy_1737122910960/work/dist/scipy-1.15.1-cp313-cp313-macosx_10_15_x86_64.whl#sha256=8d0719e8b2aa562a35f586655b39dec1277a2de1bfc1889140a803574dd1d03a
seaborn==0.13.2
setuptools==72.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.3.0
webencodings==0.5.1
wheel==0.44.0
yfinance==0.2.54
